WIP: refactor to separate agents and runners
